server/ai/requirements.txt
tiktoken
langchain_openai
python-dotenv
langchain
sentence-transformers
transformers
nltk
chromadb
langchain-community
pypdf